Restructured Flask Restful service
