+++ schema/share.py
@@ -1,4 +1,5 @@
 import re
+
 from marshmallow import Schema, fields, validates, ValidationError
